Move vision models to own submodule
